requirements.txt
aenum~=3.1.8
aiomysql~=0.1.1
aioredis~=2.0.1
AlbertUnruhUtils.py~=2.2.1
attr~=0.3.1
cryptography~=38.0.1
NAFF~=1.11.1
orjson~=3.8.0
Pillow~=9.2.0
python-dotenv~=0.21.0
PyYAML~=6.0
SQLAlchemy~=1.4.31
